requirements.txt
ipython==8.12.3
ipywidgets==8.1.7
langchain_mcp_adapters==0.1.7
langchain_openai==0.3.25
langgraph==0.4.8
prompt_toolkit==3.0.51
python-dotenv==1.1.1
rich==14.0.0
